+++ camdecmpsaux/tables/program_parameter.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS camdecmpsaux.program_parameter
     prg_id bigint NOT NULL,
     parameter_cd character varying(8) COLLATE pg_catalog."default" NOT NULL,
     required_ind numeric(1,0) NOT NULL DEFAULT 0,
-    begin_rpt_period_id bigint NOT NULL,
-    end_rpt_period_id bigint,
+    begin_rpt_period_id numeric(38,0) NOT NULL,
+    end_rpt_period_id numeric(38,0),
     userid character varying(160) COLLATE pg_catalog."default" NOT NULL,
     add_date timestamp without time zone NOT NULL,
     update_date timestamp without time zone
