+++ camdecmpswks/tables/derived_hrly_value.sql
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS camdecmpswks.derived_hrly_value
     calc_pct_moisture numeric(5,1),
     calc_rata_status character varying(75) COLLATE pg_catalog."default",
     calc_appe_status character varying(75) COLLATE pg_catalog."default",
-    rpt_period_id bigint NOT NULL,
+    rpt_period_id numeric(38,0) NOT NULL,
     mon_loc_id character varying(45) COLLATE pg_catalog."default" NOT NULL,
     calc_fuel_flow_total numeric(15,4),
     calc_hour_measure_cd character varying(7) COLLATE pg_catalog."default"
